:recycle: group module refactored to a standalone file
All group related functions refactored to group module
